fix: update example test cases to match actual code generation
- Changed integer(4) to integer in all example test cases
- Fixed function parameter declarations to use combined syntax
- All changes align with current code generator output

+++ example/frontend_test_cases/single_assignment/single_assignment.f90
@@ -1,6 +1,6 @@
 program main
     implicit none
-    integer(4) :: x
+    integer :: x
 
     x = 1
 end program main
